requirements.txt
clip==0.2.0
numpy==1.23.4
openai_clip==1.0.1
Pillow==9.4.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.64.1
